app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Document processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# ML & Embeddings
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0
pandas>=2.1.0
pyarrow>=14.0.0

# Vector DB
chromadb>=0.4.18

# LLM
llama-cpp-python>=0.2.20

# Utilities
PyYAML>=6.0.0
python-dateutil>=2.8.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Optional: Advanced OCR
# paddleocr==2.7.0.3
